pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.3",
    "bcrypt>=4.3.0",
    "celery>=5.5.3",
    "fastapi[standard]>=0.116.0",
    "gunicorn>=23.0.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pylint>=3.3.7",
    "pymysql>=1.1.1",
    "python-decouple>=3.8",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]
